Refactor: move profile components to /pages/profile. Profile header now a separate component.
