+++ src/app/App.tsx
@@ -10,7 +10,7 @@ import Search from './pages/Search'
 import Playlist from './pages/playlist/Playlist'
 import Album from './pages/album/Album'
 import Artist from './pages/artist/Artist'
-import Profile from './pages/Profile'
+import Profile from './pages/profile/Profile'
 import Account from './pages/Account'
 import Download from './pages/Download'
 import Callback from './pages/Callback'
